Update Parser to allow for multiple-word commands

+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -40,35 +40,27 @@ public class AddressBookParser {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, HelpCommand.MESSAGE_USAGE));
         }
 
-        final String commandWord = matcher.group("commandWord");
         final String arguments = matcher.group("arguments");
-        switch (commandWord) {
 
-        case AddCommand.COMMAND_WORD:
+        // using if-else and String.startsWith() instead of switch-case
+        // allows for the flexibility of having commands with multiple words, e.g. rm -contacts
+        if (userInput.startsWith(AddCommand.COMMAND_WORD)) {
             return new AddCommandParser().parse(arguments);
-
-        case EditCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(EditCommand.COMMAND_WORD)) {
             return new EditCommandParser().parse(arguments);
-
-        case DeleteCommand.COMMAND_WORD:
-            return new DeleteCommandParser().parse(arguments);
-
-        case ClearCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(ClearCommand.COMMAND_WORD)) {
             return new ClearCommand();
-
-        case FindCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(DeleteCommand.COMMAND_WORD)) {
+            return new DeleteCommandParser().parse(arguments);
+        } else if (userInput.startsWith(FindCommand.COMMAND_WORD)) {
             return new FindCommandParser().parse(arguments);
-
-        case ListCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(ListCommand.COMMAND_WORD)) {
             return new ListCommand();
-
-        case ExitCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(ExitCommand.COMMAND_WORD)) {
             return new ExitCommand();
-
-        case HelpCommand.COMMAND_WORD:
+        } else if (userInput.startsWith(HelpCommand.COMMAND_WORD)) {
             return new HelpCommand();
-
-        default:
+        } else {
             throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
         }
     }
